fix: write new test for login function
